requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==3.2.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.19.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.8
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
